Update file homework 9 task 64. Chaged type data start and end.

+++ cSharp/homeWork9_1/HomeWork9_1.cs
@@ -13,7 +13,7 @@ int start = 1;
 
 Console.WriteLine(PrintNumber(start, end));
 
-string PrintNumber(long start, long end)
+string PrintNumber(int start, int end)
 {
     if (start == end)
     {
